Renamed vcs_platform module to code_repository_platforms
Fixes #161
