+++ howfairis/mixins/license_mixin.py
@@ -1,6 +1,6 @@
 import requests
 from bs4 import BeautifulSoup
-from howfairis.vcs_platform import Platform
+from howfairis.code_repository_platforms import Platform
 
 
 class LicenseMixin:
