+++ howfairis/mixins/repository_mixin.py
@@ -1,5 +1,5 @@
 import requests
-from howfairis.vcs_platform import Platform
+from howfairis.code_repository_platforms import Platform
 
 
 class RepositoryMixin:
